# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UniversityShortResponse:
    description: Ответ API с краткими сведениями об университете.
    properties:
      like:
        description: Лайкнут ли университет пользователем
        example: true
        type: boolean
      logo:
        description: URL логотипа
        example: https://example.com/logo.png
        type: string
      name:
        description: Полное название
        example: Московский государственный университет
        type: string
      region:
        description: Название региона
        example: Москва
        type: string
      short_name:
        description: Краткое название
        example: МГУ
        type: string
      university_id:
        description: ID университета
        example: 123
        type: integer
    type: object
  errs.AppError:
    description: Структура ошибки, возвращаемая API в случае неудачного запроса.
    properties:
      code:
        description: HTTP-код ошибки
        type: integer
      details:
        additionalProperties: true
        description: Дополнительные сведения об ошибке (если есть)
        type: object
      message:
        description: Сообщение об ошибке
        type: string
      type:
        description: Тип ошибки
        type: string
    type: object
info:
  contact:
    email: olympguide@mail.ru
    name: Support Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: OlympGuide API
  version: "1.0"
paths:
...
  /olympiads/{id}/universities:
    get:
      consumes:
      - application/json
      description: Возвращает список университетов с учетом фильтров поиска и сортировкой
        по убыванию популярности.
      parameters:
      - description: ID олимпиады
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: multi
        description: Фильтр по названию регионов
        in: query
        items:
          type: string
        name: region_id
        type: array
      - description: Поиск по названию или сокращенному названию
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список университетов
          schema:
            items:
              $ref: '#/definitions/dto.UniversityShortResponse'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получение университетов, в которые олимпиада даёт льготы.
      tags:
      - Университеты
...
  /universities:
    get:
      consumes:
      - application/json
      description: Возвращает список университетов с фильтрами поиска, регионами и
        сортировкой по убыванию популярности.
      parameters:
      - collectionFormat: multi
        description: Фильтр по названию регионов
        in: query
        items:
          type: string
        name: region_id
        type: array
      - description: 'Фильтр: только университеты из региона пользователя'
        in: query
        name: from_my_region
        type: boolean
      - description: Поиск по названию или сокращенному названию
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список университетов
          schema:
            items:
              $ref: '#/definitions/dto.UniversityShortResponse'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получение списка университетов
      tags:
      - Университеты
...
  /user/diploma/{id}/universities:
    get:
      consumes:
      - application/json
      description: Возвращает список университетов с учетом фильтров поиска и сортировкой
        по убыванию популярности.
      parameters:
      - description: ID диплома
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: multi
        description: Фильтр по названию регионов
        in: query
        items:
          type: string
        name: region_id
        type: array
      - description: Поиск по названию или сокращенному названию
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список университетов
          schema:
            items:
              $ref: '#/definitions/dto.UniversityShortResponse'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получение университетов, в которые диплом даёт льготы.
      tags:
      - Университеты
...
  /user/diplomas/universities:
    get:
      consumes:
      - application/json
      description: Возвращает список университетов с учетом фильтров поиска и сортировкой
        по убыванию популярности.
      parameters:
      - collectionFormat: multi
        description: Фильтр по названию регионов
        in: query
        items:
          type: string
        name: region_id
        type: array
      - description: Поиск по названию или сокращенному названию
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список университетов
          schema:
            items:
              $ref: '#/definitions/dto.UniversityShortResponse'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получение университетов, в которые дипломы пользователя дают льготы.
      tags:
      - Университеты
swagger: "2.0"